zoteroapiserv/requirements.txt
requests>=2.28.0
python-dotenv>=0.20.0
mysql-connector-python
beautifulsoup4
lxml
feedparser
python-dateutil==2.8.2

